commit add

+++ vs2022/OglRender/OglUI/UI_SliderMultipleDoubles.h
@@ -10,12 +10,12 @@ namespace UI
 	public:
 		UI_SliderMultipleDoubles
 		(
-			float pMin = 0.0f,
-			float pMax = 1.0f,
-			float pValue = 0.5f,
+			double pMin = 0.0f,
+			double pMax = 1.0f,
+			double pValue = 0.5f,
 			const std::string& pLabel = "",
 			const std::string& pFormat = "%.6f"
-		) : UI_SliderMultipleScalars<float, _Size>(ImGuiDataType_Double, pMin, pMax, pValue, pLabel, pFormat)
+		) : UI_SliderMultipleScalars<double, _Size>(ImGuiDataType_Double, pMin, pMax, pValue, pLabel, pFormat)
 		{
 		}
 	};
